requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# mysite/crud/admin.py: 1
# mysite/crud/apps.py: 1
# mysite/crud/models.py: 1
# mysite/crud/tests.py: 1
# mysite/crud/urls.py: 1
# mysite/crud/views.py: 1,3
# mysite/manage.py: 10
# mysite/mysite/urls.py: 17
# mysite/mysite/wsgi.py: 12
Django == 2.2.2

# mysite/crud/views.py: 7,10
google_api_python_client == 1.7.9

# mysite/crud/views.py: 9
google_auth == 1.6.3

# mysite/crud/views.py: 8
google_auth_oauthlib == 0.4.0
